+++ internal/sandbox/remote/register.go
@@ -1,0 +1,7 @@
+package remote
+
+import "github.com/ash-repwiki/ash/internal/sandbox"
+
+func init() {
+	sandbox.RegisterRemoteAvailable(Available)
+}